Drop production compile tag
It does not make sense to me at this point, only makes things more complicated. We should use compilation flags only when we really need some files/packages to be not compiled - that's definitely a good case for testing (stubbing of DAO). For production, I think it is a good idea to be as close as for development environment.

This updates the Makefile and makes test flags configurable, some tests could be actually skipped with compilation flags in the future.

This now works:

* go build & run
* make build & run
* make test
