+++ internal/config/main_config.go
@@ -4,7 +4,6 @@ import (
 	"errors"
 	"fmt"
 	"regexp"
-	"strings"
 	"time"
 
 	"github.com/rs/zerolog"
@@ -67,21 +66,6 @@ var Logging = &config.Logging
 var Cloudwatch = &config.Cloudwatch
 var AWS = &config.AWS
 var Features = &config.FeatureFlags
-
-func init() {
-	viper.SetDefault("app.name", "provisioning")
-	viper.SetDefault("app.port", 8000)
-	viper.SetDefault("logging.level", 1)
-	viper.SetDefault("database.host", "localhost")
-	viper.SetDefault("database.port", 5432)
-	viper.SetDefault("database.loglevel", 1)
-	viper.SetDefault("cloudwatch.enabled", false)
-	viper.SetDefault("prometheus.port", 9000)
-	viper.SetDefault("prometheus.path", "/metrics")
-
-	viper.SetEnvKeyReplacer(strings.NewReplacer(".", "_"))
-	viper.AutomaticEnv()
-}
 
 func Initialize() {
 	if clowder.IsClowderEnabled() {
